Final Degree Project
Gas turbine blade cooling arises from the need to limit the temperature of the material, given the continous increase in the combustion temperature in order to
achieve greater efficiency.
The aim of this project is to study the effect of blade cooling on the efficiency and net specific work of a gas turbine. First, the gas turbine as a heat engine is introduced, describing its components and operation, and then various uncooled gas turbine cycles are studied. Next, a brief introduction to the state of the art of turbine blade cooling is given, firstly by studying high-temperature materials and then by studying current cooling techniques. Finally, the thermodynamic modelling of gas turbine cycles with cooling is addressed. The so-called air standard cycles is analysed first, followed by the open gas turbine cycle with cooling in the first nozzle vane row. Two of these thermodynamic
models will be implemented in Matlab software to evaluate the impact of cooling on the efficiency and net specific work.